grouping tests into basic and offchainv2 modules under suites
