+++ src/diem/testing/suites/basic/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) The Diem Core Contributors
+# SPDX-License-Identifier: Apache-2.0
