+++ src/diem/testing/suites/basic/test_generate_account_identifier.py
