+++ src/diem/testing/suites/offchainv2/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) The Diem Core Contributors
+# SPDX-License-Identifier: Apache-2.0
